+++ docs/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Next
+* Removed is:curated as an alias for is:wishlist.
 
 # 5.34.0 (2019-06-23)